requirements.txt
rank-bm25
numpy
sentence-transformers
tqdm
python-dotenv
